etl-pipeline-gcp/src/load/requirements.txt
google-cloud-storage==2.5.0
google-cloud-bigquery==2.34.0
google-cloud-functions==1.0.0
pandas==1.3.3
numpy==1.21.2
functions-framework==3.4.0
google-cloud-storage==2.9.0
google-cloud-bigquery==3.9.0
python-dotenv==1.0.0